revise data model including subclassing Cell and Command

+++ environment/mapgen/Generator.cpp
@@ -1,5 +1,5 @@
-#include "Generator.hpp"
 #include "BasicGenerator.hpp"
+#include "Generator.hpp"
 #include "Map.hpp"
 
 namespace hlt {
@@ -9,7 +9,7 @@ Generator::Generator(const MapParameters &parameters) {
     rng = std::mt19937(parameters.seed);
 }
 
-hlt::Map Generator::Map(int width, int height) const {
+hlt::Map Generator::Map(long width, long height) const {
     return hlt::Map::Map(width, height);
 }
